move mesh/rest/aws/clients to /pkg/compute-target/aws
